WIP: pre-Orca-port snapshot of in-flight work
Captures uncommitted work prior to forking off to import Orca's
tree-support engine on a separate branch. Includes:

- BP1007 cleanup: removed redundant 'using namespace' directives in
  SparseLineGrid.hpp, ArrangeSettingsView.hpp, and
  SVGDebugOutputKernelWrapper.hpp
- VS 2026 / MSVC 14.50 build adaptations across vendored deps
  (admesh, angelscript, eigen, libigl, libnest2d, hidapi,
  boost::nowide, glu-libtess)
- Tree-support investigation work (Support/OrganicSupport.cpp,
  PrintConfig)
- Various GUI, GCode, Format, SLA, and Config in-progress edits

+++ src/libigl/igl/copyleft/cgal/order_facets_around_edge.cpp
@@ -323,7 +323,7 @@ void igl::copyleft::cgal::order_facets_around_edge(
 
   // Because face indices are used for tie breaking, the original face indices
   // in the new faces array must be ascending.
-  auto comp = [&](int i, int j) 
+  auto comp = [&](size_t i, size_t j) 
   {
     return signed_index_to_index(adj_faces[i]) <
       signed_index_to_index(adj_faces[j]);
